technical-tests/age-bmi/requirements.txt
python-dotenv==1.1.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pandas==2.3.2
streamlit==1.49.1
fhirclient==4.3.2
rcpchgrowth==4.0.0
Pint==0.25
pytest==8.4.2
testcontainers==4.13.0